+++ src/test/java/com/github/danielfabela/partpicker/math/ComponentsTest.java
@@ -1,8 +1,0 @@
-package com.github.danielfabela.partpicker.math;
-
-import org.junit.Before;
-import org.junit.Test;
-
-public class ComponentsTest {
-
-}